backend/requirements.txt
annotated-types
anyio
blinker
certifi
click
distro
docx
h11
httpcore
httpx
idna
itsdangerous
Jinja2
jiter
lxml
MarkupSafe
openai
pillow
pydantic_core
PyPDF2
python-docx
python-dotenv
sniffio
tqdm
typing_extensions
Werkzeug
fastapi
uvicorn
motor
beanie
pydantic
passlib
python-jose
python-multipart

